+++ src/main/java/com/xavelo/crypto/service/RateConversionService.java
@@ -1,12 +1,10 @@
 package com.xavelo.crypto.service;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Service;
 
 import com.xavelo.crypto.data.Currency;
 
 @Service
 public interface RateConversionService {    
-    BigDecimal convertToUSD(Currency from, BigDecimal amount);
+    double convertToUSD(Currency from, double amount);
 }
